blink: Enable -Wshorten-64-to-32 in renderer/platform/
Add "//third_party/blink/renderer:narrowing" to the platform targets. It
contains -Wtautological-unsigned-zero-compare too, so this CL fixes
warning by it.

We need to fix truncation issues in media/base/audio_bus.h and
media/base/sinc_resampler.h because they are included from blink code.

Bug: 514751964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7894702
Commit-Queue: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1642094}

+++ media/base/audio_bus.h
@@ -18,6 +18,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/memory/aligned_memory.h"
 #include "base/memory/raw_ptr_exclusion.h"
 #include "base/numerics/checked_math.h"
+#include "base/numerics/safe_conversions.h"
 #include "media/base/media_export.h"
 
 namespace media {
@@ -204,7 +205,7 @@ class MEDIA_EXPORT AudioBus {
   // Note: for bitstream formats, use GetBitstreamFrames() to get the actual
   // number of encoded frames. However, `frames()` remains useful in determining
   // the amount of `reserved_memory_` this bus has.
-  int frames() const { return frames_; }
+  int frames() const { return base::checked_cast<int>(frames_); }
 
   // Helper method for zeroing out all channels of audio data.
   void Zero();
@@ -312,7 +313,8 @@ void AudioBus::FromInterleaved(
 
   // If not using `zero_remaining_frames`, `source` should have the exact size.
   CHECK(zero_remaining_frames);
-  ZeroFramesPartial(source_frame_count, remaining_frames);
+  ZeroFramesPartial(base::checked_cast<int>(source_frame_count),
+                    base::checked_cast<int>(remaining_frames));
 }
 
 template <class SourceSampleTypeTraits>
@@ -361,7 +363,8 @@ void AudioBus::CopyConvertFromInterleavedSourceToAudioBus(
   CHECK_LE(total_offset, static_cast<size_t>(dest->frames()));
 
   for (size_t ch = 0; ch < channels; ++ch) {
-    auto channel_data = dest->channel(ch).subspan(write_offset, frame_count);
+    auto channel_data =
+        dest->channel(static_cast<int>(ch)).subspan(write_offset, frame_count);
     for (size_t dest_idx = 0, src_idx = ch; dest_idx < frame_count;
          ++dest_idx, src_idx += channels) {
       auto source_sample = source[src_idx];
@@ -382,7 +385,8 @@ void AudioBus::CopyConvertFromAudioBusToInterleavedTarget(
   CHECK_LE(total_offset, static_cast<size_t>(source->frames()));
 
   for (size_t ch = 0; ch < channels; ++ch) {
-    auto channel_data = source->channel(ch).subspan(read_offset, frame_count);
+    auto channel_data =
+        source->channel(static_cast<int>(ch)).subspan(read_offset, frame_count);
     for (size_t src_idx = 0, dest_idx = ch; src_idx < frame_count;
          ++src_idx, dest_idx += channels) {
       float source_sample = channel_data[src_idx];
