+++ media/base/sinc_resampler.h
@@ -14,6 +14,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/memory/aligned_memory.h"
 #include "base/memory/raw_ptr.h"
 #include "base/memory/raw_span.h"
+#include "base/numerics/safe_conversions.h"
 #include "build/build_config.h"
 #include "media/base/media_export.h"
 
@@ -73,7 +74,7 @@ class MEDIA_EXPORT SincResampler {
   // a single call to |read_cb_| for more data.  Note: If PrimeWithSilence() is
   // not called, chunk size will grow after the first two Resample() calls by
   // `kernel_size_` / (2 * io_sample_rate_ratio).  See the .cc file for details.
-  int ChunkSize() const { return chunk_size_; }
+  int ChunkSize() const { return base::checked_cast<int>(chunk_size_); }
 
   // Returns the max number of frames that could be requested (via multiple
   // calls to |read_cb_|) during one Resample(|output_frames_requested|) call.
